Added compatability for Doctrine Common 2.1
Squashed commit of the following:

commit 14403a75dcc8b910e1881997451d67c2a4a8e555
Author: Richard Fullmer <rfullmer@printfulfillmentservices.com>
Date:   Fri Jul 15 11:25:46 2011 -0400

    Added compatability for doctrine common 2.1

commit 4d83afc422fc94401a20e76c1807f7415c153698
Author: Richard Fullmer <rfullmer@printfulfillmentservices.com>
Date:   Fri Jul 15 09:55:07 2011 -0400

    Fixing a namespace issue with the DoctrineAnnotations

commit 2fcc3712cd10998445ec7925f6f7a8eb890ddb23
Author: Richard Fullmer <rfullmer@printfulfillmentservices.com>
Date:   Fri Jul 15 09:52:45 2011 -0400

    Updating doctrine-common to 2.1 release
