+++ tests/Doctrine/Tests/OXM/Mapping/AbstractMappingDriverTest.php
@@ -198,110 +198,3 @@ abstract class AbstractMappingDriverTest extends \Doctrine\Tests\OxmTestCase
         return $class;
     }
 }
-
-/**
- * @XmlRootEntity(xml="cms-user")
- * @HasLifecycleCallbacks
- * @XmlNamespaces({
- *   @XmlNamespace(url="http://www.schema.com/foo", prefix="foo"),
- *   @XmlNamespace(url="http://www.schema.com/bar", prefix="bar")
- * })
- */
-class User
-{
-    /**
-     * @XmlId
-     * @XmlField(type="string", node="attribute")
-     */
-    public $id;
-
-    /**
-     * @XmlText(type="string", required=true, getMethod="getUsername", setMethod="setUsername")
-     */
-    public $name;
-
-    /**
-     * @XmlText(type="string", collection=true, wrapper="comments", name="comment")
-     */
-    public $comments;
-
-    /**
-     * @PrePersist
-     */
-    public function doStuffOnPrePersist()
-    {
-
-    }
-    /**
-     * @PrePersist
-     */
-    public function doOtherStuffOnPrePersistToo()
-    {
-
-    }
-
-    /**
-     * @PostPersist
-     */
-    public function doStuffOnPostPersist()
-    {
-        
-    }
-
-    /**
-     * @PreMarshal
-     */
-    public function doStuffOnPreMarshal()
-    {
-
-    }
-
-    public function getUsername()
-    {
-        return $this->name;
-    }
-
-    public function setUsername($name)
-    {
-        $this->name = $name;
-    }
-
-    public static function loadMetadata(ClassMetadataInfo $metadata)
-    {
-        $metadata->setXmlName('cms-user');
-        $metadata->isRoot = true;
-
-        $metadata->setXmlNamespaces(array(
-            array('url' => 'http://www.schema.com/foo', 'prefix' => 'foo'),
-            array('url' => 'http://www.schema.com/bar', 'prefix' => 'bar')
-        ));
-
-        $metadata->addLifecycleCallback('doStuffOnPrePersist', 'prePersist');
-        $metadata->addLifecycleCallback('doOtherStuffOnPrePersistToo', 'prePersist');
-        $metadata->addLifecycleCallback('doStuffOnPostPersist', 'postPersist');
-        $metadata->addLifecycleCallback('doStuffOnPreMarshal', 'preMarshal');
-
-        $metadata->mapField(array(
-            'fieldName' => 'id',
-            'id' => true,
-            'type' => 'string',
-            'node' => 'attribute'
-        ));
-        $metadata->mapField(array(
-            'fieldName' => 'name',
-            'type' => 'string',
-            'node' => 'text',
-            'required' => true,
-            'setMethod' => 'setUsername',
-            'getMethod' => 'getUsername'
-        ));
-        $metadata->mapField(array(
-            'fieldName' => 'comments',
-            'type' => 'string',
-            'node' => 'text',
-            'collection' => true,
-            'wrapper' => 'comments',
-            'name' => 'comment',
-        ));
-    }
-}
