+++ tests/Doctrine/Tests/OXM/Mapping/AnnotationDriverTest.php
@@ -19,6 +19,7 @@
 
 namespace Doctrine\Tests\OXM\Mapping;
 
+use Doctrine\Common\Cache\ArrayCache;
 use Doctrine\OXM\Mapping\ClassMetadata;
 use Doctrine\OXM\Mapping\ClassMetadataInfo;
 use Doctrine\OXM\Mapping\Driver\Driver;
@@ -30,10 +31,23 @@ class AnnotationDriverTest extends AbstractMappingDriverTest
      */
     protected function _loadDriver()
     {
-        $cache = new \Doctrine\Common\Cache\ArrayCache();
-        $reader = new \Doctrine\Common\Annotations\AnnotationReader($cache);
-        $reader->setDefaultAnnotationNamespace('Doctrine\OXM\Mapping\\');
-        return new \Doctrine\OXM\Mapping\Driver\AnnotationDriver($reader);
+        if (version_compare(\Doctrine\Common\Version::VERSION, '3.0.0', '>=')) {
+            $reader = new \Doctrine\Common\Annotations\CachedReader(
+                new \Doctrine\Common\Annotations\AnnotationReader(), new ArrayCache()
+            );
+        } else if (version_compare(\Doctrine\Common\Version::VERSION, '2.1.0-BETA3-DEV', '>=')) {
+            $reader = new \Doctrine\Common\Annotations\AnnotationReader();
+            $reader->setIgnoreNotImportedAnnotations(true);
+            $reader->setEnableParsePhpImports(false);
+                $reader->setDefaultAnnotationNamespace('Doctrine\OXM\Mapping\\');
+            $reader = new \Doctrine\Common\Annotations\CachedReader(
+                new \Doctrine\Common\Annotations\IndexedReader($reader), new ArrayCache()
+            );
+        } else {
+            $reader = new \Doctrine\Common\Annotations\AnnotationReader();
+            $reader->setDefaultAnnotationNamespace('Doctrine\OXM\Mapping\\');
+        }
+        return new \Doctrine\OXM\Mapping\Driver\AnnotationDriver($reader, array(__DIR__));
     }
 
 }
