+++ tests/Doctrine/Tests/OXM/Marshaller/MarshallerTypeTest.php
@@ -35,7 +35,7 @@ class MarshallerTypeTest extends \PHPUnit_Framework_TestCase
     public function setUp()
     {
         $config = new Configuration();
-        $config->setMetadataDriverImpl(AnnotationDriver::create("tests/Doctrine/Tests/OXM/Marshaller"));
+        $config->setMetadataDriverImpl($config->newDefaultAnnotationDriver("tests/Doctrine/Tests/OXM/Marshaller"));
         $config->setMetadataCacheImpl(new \Doctrine\Common\Cache\ArrayCache());
 
         $metadataFactory = new ClassMetadataFactory($config);
